Fix cache invalidation to update tex every build
Bit of a hack, but quite neat! `RUN` is cached and never invalidated,
but `ADD <url>` always runs, and invalidates all that follows iff the
response differs to the previously cached response.

So this commit uses a URL which will 'always' differ in order to force
the invalidation of the cache for the next line, (and any that followed,
if there were) meaning that the `RUN` to update TeX packages will always
run on every build, even if it's been previously cached.